backend/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
httpx==0.27.0
python-jose==3.3.0
itsdangerous==2.2.0
pydantic==1.10.15
openai==1.35.7
SQLAlchemy==2.0.30
asyncpg==0.29.0
PokerKit==0.6.5
